% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbackConnection.R
\name{callbackConnection}
\alias{callbackConnection}
\title{Create a connection which executes an R function upon receiving a message}
\usage{
callbackConnection(callback = function(char) print(char))
}
\arguments{
\item{callback}{an R function to execute with each message}
}
\value{
a new callbackConnection connection object
}
\description{
This connection can be provided with an arbitrary R callback function that
will be executed silently and without producing additional output for each
new messag received by the connection.
}
